pyproject.toml
[project]
name = "langchain-agents-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.27",
    "langchain-openai>=0.3.28",
    "langsmith>=0.4.13",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
]
